% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cartograflow}
\alias{cartograflow}
\alias{cartograflow-package}
\title{Cartograflow Package}
\description{
This package is designed to create the so-called flowmaps, by filtering origin-destination (OD) matrix.
It is based on different functions that are mainly used to prepare the flow dataset.
The spatial objects processing are those of {sp} or {sf} and
the mapping elements are often those of {Cartography} except particular cases.
}
\details{
To learn more about cartograflow, see the vignette cartograflow.html

Main functions :
\itemize{
 \item{
   \link{flowanalysis} \link{flowcarre} \link{flowcontig} \link{flowdist}
   \link{flowgini} \link{flowjointure}  \link{flowmap} \link{flowreduct}
   \link{flowstructmat} \link{flowcontig}
      }
 }
}
