% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowanalysis.R
\name{flowanalysis}
\alias{flowanalysis}
\title{Computation of a global selection criterion for flowmapping}
\usage{
flowanalysis(tab, critflow, critlink, result)
}
\arguments{
\item{tab}{flow dataset from \link{flowgini}}

\item{critflow}{level of flow significativity. See Details.}

\item{critlink}{level of features density. See Details.}

\item{result}{resulting criterion value. See Details}
}
\description{
Computation of a global selection criterion for thresholding flow  information or features before mapping
To be use after \link{flowgini}
}
\details{
Computes a double criterion for selecting flows before mapping :
-- critflow =  desired level of flow's information significativity, as a % of total of flow information ;
-- critlink = desired level of flow's features density, as a % of total of flow features.

Return resuts :
-- result="density" returns the desired level of features density as a % of total features ;
-- result = "significativity" returns the level of flow significativity as a % of total of flow information ;
}
\references{
Bahoken Francoise (2016), « La cartographie d’une sélection globale de flux, entre ‘significativité’ et ‘densité’ »,
Netcom [On ligne], 30-3/4 | URL : http://journals.openedition.org/netcom/2565 ; DOI : 10.4000/netcom.2565
Bahoken Francoise (2016), « Chapitre 8 Propositions de solutions liées au choix d’un critère de
sélection global : la cartographie de flux significatifs (Fij>α)» in Bahoken, F. Contribution à la cartographie d'une matrice de flux, Thèse de doctorant, Université Paris 7, pp. 325-346.
Grasland Claude (2011, 2014), « Flows analysis carto », unpublished R functions.
}
