% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcontig.R
\name{flowcontig}
\alias{flowcontig}
\title{Computes ordinal distance matrixes based on contiguity}
\usage{
flowcontig(fdc, code, ordre)
}
\arguments{
\item{fdc}{is the map background file, ie. a shapefile of polygons}

\item{code}{identifiant in shape file}

\item{ordre}{ordre est à 1, 2 ou 4}
}
\value{
a (k) contiguity matrice with the k contiguity measures
}
\description{
From a spatial basemap, compute an ordinal distance matrix based on a k-contiguity matrix.
The resulting neighbourhood graph can be map using \link{flowmap}
}
\details{
Contiguity is in terms of the (k=1,2,4) numberof spatial boundaries to be crossed between
a place of origin and a place of destination
ordre=1 is when the flow have to cross only 1 boundary, that is to say the corresponding origin and destination places are
contiguous, adjacent
ordre=2 is when the origin-destinations places are distant from 2 borders
ordre=4 is when the origin-destinations places are distant from 4 borders
}
