% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowdist.R
\name{flowdist}
\alias{flowdist}
\title{Compute continous distance matrice from geographical background}
\usage{
flowdist(tab, dist.method, result)
}
\arguments{
\item{tab}{the input flow dataset.}

\item{dist.method}{euclidian calculation}

\item{result}{take the vallue "flowdist" or "dist"allows to parameter the resulting distance dataset (flows filtered by a distance criterion or not)}
}
\value{
(1) A flowdata set with continuous euclidian distances calculations, see dist.method parameter

(2) A flowdata set with movement from euclidian distances calculations.

(3) A flowmap filtered by a global distance criterion.
}
\description{
From a geographical backhround computes (and threshold) a distance matrice.
}
\details{
-- result = "flowdist" is tab with all calculation
-- result = "dist" is tab of the distance
}
