% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowjointure.R
\name{flowjointure}
\alias{flowjointure}
\title{Create a spatial joint between a flow dataset table and a map background}
\usage{
flowjointure(tab, fdc, code)
}
\arguments{
\item{tab}{the input flow dataset table in long format}

\item{fdc}{the map background file, ie. a shapefile}

\item{code}{the ID of the spatial units in the map background}
}
\value{
Resulting jointure table between flow dataset and map background
}
\description{
Performs an attribute spatial join between a flow dataset table and a map background
}
\examples{
\dontrun{
library(cartograflow)
data<-read.csv2("flowdataset.csv",
                 header=TRUE,
                 sep=";",
                 stringsAsFactors=FALSE,
                 encoding="UTF-8",
                 dec=".",
                 check.names=FALSE)
tab<-data \%>\% select(i,j,Fij)
tabflow<-flowjointure(tab,"bkg.shp","code")
}
}
