% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowstructmat.R
\name{flowstructmat}
\alias{flowstructmat}
\title{Structuring a matrice}
\usage{
flowstructmat(z)
}
\arguments{
\item{z}{The flow dataset is in the matrice format where the first column is filled with the ID}
}
\value{
A flowdataset with an usable format
}
\description{
Fixes an ID shift in the flow  matrice (to bo use with \link{flowjointure} if necessary and \link{flowtabmat})
}
\examples{
library(cartograflow)
data(flowdata)

dim(mat_ex) # dimension fo the original matrice
### 10 11  # first colum is fill with the ID

tab<-flowstructmat(mat_ex)
dim(tab)
## 10 10 # dimension fo the resulting matrice
}
