% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geoid}
\alias{geoid}
\title{Geographical ID}
\source{
\url{https://www.insee.fr/fr/statistiques/fichier/3566008/rp2015_mobpro_txt.zip}
}
\description{
\code{One column dataframe in.csv.}\cr
\code{Variable (COD_GEO_EPT) is the geographical code of the territory}\cr
\code{citation : APUR, 2018}
}
