% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cartograflow}
\alias{cartograflow}
\title{Cartograflow Package}
\description{
This package contains functions to prepare and filter origin-destination matrix
for thematic flow mapping purposes. The spatial objects processing are those of {sf}.\cr
This comes after Bahoken, Francoise (2016) Contribution à la cartographie d'une matrice de flux, 
Phd in Geography, Sorbonne Paris Cité, Paris 7.
}
\details{
To learn more about cartograflow, see the vignette cartograflow.html

Main functions :
\itemize{
 \item{
   \link{flowanalysis} \link{flowcarre} \link{flowcontig} \link{flowdist}
   \link{flowgini} \link{flowjointure}  \link{flowmap} \link{flowreduct}
   \link{flowstructmat} \link{flowcontig}
      }
 }
}
