% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowanalysis.R
\name{flowanalysis}
\alias{flowanalysis}
\title{Computation of a global selection criterion for thresholding flow values and/or flow features}
\usage{
flowanalysis(tab, critflow, critlink, result)
}
\arguments{
\item{tab}{flow dataset from \link{flowgini}}

\item{critflow}{level of flow significativity. See Details.}

\item{critlink}{level of features density. See Details.}

\item{result}{resulting filtering criterion value. See Details}
}
\description{
To be use after \link{flowgini}.
Computation of a global selection criterion for thresholding flow information and/or features before mapping.
}
\details{
-- critflow =  desired level of flow's information significativity (e.g. 80% of the total information) ;
-- critlink = desired level of flow's features density (e.g. 20% of the flow features that represents "more significant information).

-- result="density" returns the desired level of features density as a % of total features ;
-- result = "significativity" returns the level of flow significativity as a % of total of flow information ;
}
