% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcarre.R
\name{flowcarre}
\alias{flowcarre}
\title{Create a square matrice from geographical ID}
\usage{
flowcarre(liste, tab, origin, dest, valflow, format, diagonale, empty.sq)
}
\arguments{
\item{liste}{list or all the geographical ID as a single vector flow dataset}

\item{tab}{the non squared input flow dataset with three column : origin, destination, flow value}

\item{origin}{origin place ID}

\item{dest}{destination place ID}

\item{valflow}{flow value}

\item{format}{the squared flow dataset output format : see details}

\item{diagonale}{see details}

\item{empty.sq}{TRUE : to allows to have an matrice empty with only the ID of background map ; FALSE or missing}
}
\description{
Create a closed and square matrice from a list of geographic ID (CODE)
}
\details{
- format is "M" for matrice format
- format is "L" for long format
- diagonal is TRUE to zero main diagonal
}
\examples{
\dontrun{
transform a non squared flow dataset to a closed and squared matrice
library(cartograflow)
code<-read.csv2("file1.csv",header=TRUE,sep=";",stringsAsFactors=FALSE,
encoding="UTF-8",dec=".",check.names=FALSE)
flow<-read.csv2("file2.csv",header=TRUE,sep=";",stringsAsFactors=FALSE,
                 encoding="UTF-8",dec=".",check.names=FALSE)
mat<-flowcarre(liste="code",tab="flow",origin="COMMUNE_O",dest="COMMUNE_D",
               valflow="COUNT",format="L")
}
}
