% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcarre.R
\name{flowcarre}
\alias{flowcarre}
\title{Create a square matrice from IDs}
\usage{
flowcarre(liste, tab, origin, dest, valflow, format, diagonale, empty.sq)
}
\arguments{
\item{liste}{list or all the spatial units IDs, on a single column}

\item{tab}{is the non squared input flowdata table with three column : origin, destination, flow)}

\item{origin}{identify the origin nodes column (which contains IDs)}

\item{dest}{identify the destination nodes column (which contains IDs)}

\item{valflow}{identify tne flow value column}

\item{format}{indicates the squared flow data output format : "M" or "L"for matrice (n*n) or "L" for list}

\item{diagonale}{tttt}

\item{empty.sq}{TRUE : to allows to have an matrice empty with only the ID of background map ; FALSE or missing}
}
\description{
Create a closed square matrice from a list of geographic IDs (CODE)
}
\details{
The table listing all the spatial geometry IDs involved (origin and/or destination)
must have been prepared in advance.
For a matrice (n*n) output form: format ="M"
For a list output form: format ="L"
}
\examples{
## Example 1 : transform a non squared flowdata matrice to a closed and squared matrice
#library(cartograflow)
#code<-read.csv2("file1.csv",header=TRUE,sep=";",stringsAsFactors=FALSE,
#encoding="UTF-8",dec=".",check.names=FALSE)
#flow<-read.csv2("file2.csv",header=TRUE,sep=";",stringsAsFactors=FALSE,
#encoding="UTF-8",dec=".",check.names=FALSE)
#mat<-flowcarre(tab="file1",tab="file2",origin="COMMUNE_O",dest="COMMUNE_D",
#valflow="COUNT",format="L")
}
