% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowjointure.R
\name{flowjointure}
\alias{flowjointure}
\title{Create a spatial join with flow}
\usage{
flowjointure(tab, fdc, code)
}
\arguments{
\item{tab}{the input flow dataset table in long format}

\item{fdc}{the map background file, ie. a shapefile}

\item{code}{the ID of the spatial units in the map background}
}
\value{
Resulting jointure table between flow dataset and map background
}
\description{
Create an attribute spatial join between a flow dataset table and a map background
}
\examples{
library(cartograflow)
data(flowdata)
bkg<- system.file("shape/MGP_TER.shp", package="cartograflow",
                   lib.loc = NULL, mustWork = TRUE)
tabflow<-flowjointure(MOBPRO_ETP,bkg,"EPT_NUM")
}
