% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmap.R
\name{flowmap}
\alias{flowmap}
\title{Plot a layer of origin-destination flows values}
\usage{
flowmap(tab, format, fdc, code, filter, threshold, taille, a.head,
  a.length, a.angle, a.col)
}
\arguments{
\item{tab}{is the input flowdata table}

\item{format}{is the flowdata table format : M=matrix [n*n] or L=list [, i,j, Fij].}

\item{fdc}{is the geometry or the geographical background file .shp composed of spatial units}

\item{code}{is the column with the spatial units ID code}

\item{filter}{allows you to filter (or not) the flow dataset.}

\item{threshold}{is the value of the threshold criterion used to filter the values. The default threshold is set to 1}

\item{taille}{is a graphical parameter for modifying the width of the flow feature}

\item{a.head}{integer code, determining the kind of arrows to be drawn.}

\item{a.length}{length of the edges of the arrow head (in inches).}

\item{a.angle}{angle from the shaft of the arrow to the edge of the arrow head.}

\item{a.col}{color of the arrows}
}
\value{
a matrix or a list with the correct tabflow ID code

The resulting flowmap
}
\description{
Plot a layer of origin-destination flows values
}
\details{
The flow dataset must be converted to a dataframe for optimal performance (troubles remains with tibble format)

filter allows to choose theoretical or empirical links to draw.
If filter = FALSE, all the matrice values are plot [(n*(n-1)] cells, i.e. all links out of the main diagonal.
If filter = TRUE only non-zero values are plotted, i.e. existing links with or without threshold.
The default threshold is set to 1.

varf = "True" draw arrows, unless it draw links

If varf="true", an arrows parameter is set.
code=0 : the link has no head.
code=1 : an arrow is draw at (x0[i], y0[i]).
code=2 : an arrow is draw at (x1[j], y1[j])
code=3 : an arrow is draw at both nodes.
}
\examples{
## Example 1 : flowmapping a network - draw non valued links
#library(cartograflow)
#flowmap(tabflow,format = "L",filter = FALSE, carte = "file.shp",code = "geoID",varf = FALSE)
## Example 2 : flowmapping - draw valued and oriented links or arrows
#library(cartograflow)
#flowmap(tabflow,format = "L",fdc = "file.shp",filter = TRUE,
#threshold = (user value),taille = (user value),code = "geoID",
#a.head = 1 ,a.length = 0.2,a.angle=45,a.col="blue")
}
\seealso{
\code{arrows}
}
