% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowreduct.R
\name{flowreduct}
\alias{flowreduct}
\title{Flow matrice reduction according to another matrice}
\usage{
flowreduct(tab, tab.metric, metric, select, d)
}
\arguments{
\item{tab}{is the input flowdata set.}

\item{tab.metric}{is the table of distance (continuous dataset) or contiguity (ordinal dataset)}

\item{metric}{See Details.}

\item{select}{is the continuous distance thresholding parameter. See Details.}

\item{d}{distance thresholds criterion.}
}
\value{
A flow dataset with continuous euclidian distances calculations
}
\description{
Reduces a flow dataset according to an external matrice (eg. distance travelled)
Computes geographical movements (by weighting a flow dataset according to a distance criterion)
}
\details{
This function allows to reduce a flow dataset from a matrice distance.
-- Metric is 'continous" for distance as euclidian, maximum, manhattan, etc.
See \link{flowdist} for computing neighbourhood ordinal distance matrix.

select ="dmin" is for reducing flow dataset to values that are up or equal to the dmin distance parameter.
select ="dmax" is for reducing flow dataset to values that are less or equal to the dmax distance.

Metric is 'ordinal' for neighbourhood ordinal distance so-called k contiguity.
See \link{flowcontig} for computing continuous distance matrix
}
