% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flows}
\alias{flows}
\title{MOBPRO: Commuting trips in 2015}
\source{
\url{https://www.insee.fr/fr/statistiques/fichier/3566008/rp2015_mobpro_txt.zip}
}
\description{
Statistical dataset in .csv:
\code{Extraction of the french national census : "Mobilités professionnelles en 2015 : déplacements domicile - lieu de travail"
- Base flux de mobilité - for the Greater Paris area.}\cr
\code{Citation: INSEE - RP MOBPRO, 2015.}\cr
\code{Variable (i) is the place of origin of the flow.}\cr
\code{Variable (j) is the place of destination of the flow.}\cr
\code{Variable (Fij) is the flow value between (i,j).}\cr
\code{Variable (count) is the frequency of the (i,j) couple of places.}\cr
}
