% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowtabmat.R
\name{flowtabmat}
\alias{flowtabmat}
\title{Transform a matrice format to a long format and vice versa}
\usage{
flowtabmat(tab, matlist)
}
\arguments{
\item{tab}{dasaset matrice or list}

\item{matlist}{choise of matice or list}
}
\value{
matrice or list
}
\description{
Transform a matrice format to a long format and vice versa
}
\details{
From list to matrice: matlist="M" ;
From matrice to list: matlist="L".
}
\examples{
## Example 1 : from list to matrix (n*m)
#library(cartograflow)
#tabFlow<-read.csv2("flowdataset.csv",header=TRUE,sep=";",stringsAsFactors=FALSE,
#encoding="UTF-8",dec=".",check.names=FALSE)
#matFlow<-flowtabmat(tabFlow,matlist="M")
## Example 2 : from matrix to list [i,j,Fij]
#library(cartograflow)
#matFlow<-read.csv2("flowdataset.csv",header=TRUE,sep=";",stringsAsFactors=FALSE,
#encoding="UTF-8",dec=".",check.names=FALSE)
#tabFlow<-flowtabmat(matFlow,matlist="L")
}
