% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowtabmat.R
\name{flowtabmat}
\alias{flowtabmat}
\title{Changing the format of a flow dataset}
\usage{
flowtabmat(tab, matlist)
}
\arguments{
\item{tab}{flow dasaset, in matrice or long format}

\item{matlist}{choose matrice or long as the result format. See Details.}
}
\value{
flow data in matrice or long format
}
\description{
Transform a flow dataset from long to matrice format, and vice versa
}
\details{
-- From long to matrice format [n*n]: matlist="M";\cr
-- From matrice to long format [i,j,Fij]: matlist="L".
}
\examples{
library(cartograflow)
data(flowdata)
#1: From long to matrix format (n*m)
matFlow<-flowtabmat(flows,matlist="M")
#2: From matrix to long format [i,j,Fij]
listflow<-flowtabmat(matFlow,matlist="L")
}
